Agregar script de Inno Setup para la instalación de GardataSys
- Creado setup.iss para la aplicación GardataSys
- Configurados los detalles de la aplicación incluyendo nombre, versión y editor
- Especificado el directorio de instalación y la configuración de salida
- Incluidos los archivos y directorios necesarios para la instalación
- Añadida opción de creación de icono en el escritorio
- Establecidos permisos para el directorio de la aplicación
- Configurada la ejecución posterior a la instalación de la aplicación principal

+++ src/ModeloFinal/ProductosDAO.java
@@ -186,7 +186,7 @@ public class ProductosDAO {
                 prod.setNombre(rs.getString("nombre"));
                 prod.setMedida(rs.getString("medida"));
                 prod.setCantidad(rs.getDouble("cantidad"));
-                prod.setPrecio(rs.getLong("precio"));
+                prod.setPrecio(rs.getDouble("precio"));
                 prod.setProveedor(rs.getString("proveedor"));
                 ListaPROD.add(prod);
             }
